edk2/MdeModulePkg/Universal/PcatRealTimeClockRuntimeDxe/PcRtc.c:   Add type casting to fix warning generated by the Intel compiler.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@5187 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Universal/PcatRealTimeClockRuntimeDxe/PcRtc.c
@@ -179,7 +179,7 @@ Returns:
   // Set RTC configuration after get original time
   // The value of bit AIE should be reserved.
   //
-  RtcWrite (RTC_ADDRESS_REGISTER_B, RTC_INIT_REGISTER_B | (RegisterB.Data & BIT5));
+  RtcWrite (RTC_ADDRESS_REGISTER_B, (UINT8)(RTC_INIT_REGISTER_B | (RegisterB.Data & BIT5)));
 
   //
   // Release RTC Lock.
